Make explicit how new directive inherits from CanDirective
