+++ api/graphql/schema.graphql
@@ -28,16 +28,17 @@ enum Role {
 }
 
 type User {
+    # There are some cases where some fields, such as email and names, should be visible while others aren't. Therefore we use field-level permissions to distinguish them.
     id: ID!
-    email: Email! @parentCan(ability: "viewPublic")
-    sub: String @parentCan(ability: "view")
-    firstName: String @rename(attribute: "first_name") @parentCan(ability: "viewPublic")
-    lastName: String @rename(attribute: "last_name") @parentCan(ability: "viewPublic")
-    telephone: PhoneNumber  @parentCan(ability: "view")
-    preferredLang: Language @rename(attribute: "preferred_lang") @parentCan(ability: "view")
-    roles: [Role] @parentCan(ability: "view")
-    pools: [Pool] @hasMany @parentCan(ability: "view") # Pools that a user owns
-    poolCandidates: [PoolCandidate] @hasMany @parentCan(ability: "view") # Pools candidates that a user owns
+    email: Email! @canAsParent(ability: "viewPublic")
+    sub: String @canAsParent(ability: "view")
+    firstName: String @rename(attribute: "first_name") @canAsParent(ability: "viewPublic")
+    lastName: String @rename(attribute: "last_name") @canAsParent(ability: "viewPublic")
+    telephone: PhoneNumber  @canAsParent(ability: "view")
+    preferredLang: Language @rename(attribute: "preferred_lang") @canAsParent(ability: "view")
+    roles: [Role] @canAsParent(ability: "view")
+    pools: [Pool] @hasMany @canAsParent(ability: "view") # Pools that a user owns
+    poolCandidates: [PoolCandidate] @hasMany @canAsParent(ability: "view") # Pools candidates that a user owns
 }
 
 type PoolOwnerPublicProfile {
